api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
sqlalchemy==2.0.23
pymysql==1.1.0
jinja2==3.1.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
pydantic[email]==2.5.0
aiofiles==23.2.1
fpdf2==2.7.9

# Testing
pytest==7.4.4
pytest-cov==4.1.0
httpx==0.25.2